events: extract `EventEmitterAsyncResource` to a separate file
