+++ lib/events.js
@@ -27,7 +27,6 @@ const {
   ArrayPrototypePush,
   ArrayPrototypeSlice,
   ArrayPrototypeSplice,
-  ArrayPrototypeUnshift,
   Boolean,
   Error,
   ErrorCaptureStackTrace,
@@ -68,7 +67,6 @@ const {
   AbortError,
   codes: {
     ERR_INVALID_ARG_TYPE,
-    ERR_INVALID_THIS,
     ERR_UNHANDLED_ERROR,
   },
   genericNodeError,
@@ -82,7 +80,6 @@ const {
   validateFunction,
   validateNumber,
   validateObject,
-  validateString,
 } = require('internal/validators');
 const { addAbortListener } = require('internal/events/abort_listener');
 
@@ -102,114 +99,7 @@ let EventEmitterAsyncResource;
 // eventemitter-asyncresource MIT-licensed userland module.
 // https://github.com/addaleax/eventemitter-asyncresource
 function lazyEventEmitterAsyncResource() {
-  if (EventEmitterAsyncResource === undefined) {
-    const {
-      AsyncResource,
-    } = require('async_hooks');
-
-    const kEventEmitter = Symbol('kEventEmitter');
-    const kAsyncResource = Symbol('kAsyncResource');
-    class EventEmitterReferencingAsyncResource extends AsyncResource {
-      /**
-       * @param {EventEmitter} ee
-       * @param {string} [type]
-       * @param {{
-       *   triggerAsyncId?: number,
-       *   requireManualDestroy?: boolean,
-       * }} [options]
-       */
-      constructor(ee, type, options) {
-        super(type, options);
-        this[kEventEmitter] = ee;
-      }
-
-      /**
-       * @type {EventEmitter}
-       */
-      get eventEmitter() {
-        if (this[kEventEmitter] === undefined)
-          throw new ERR_INVALID_THIS('EventEmitterReferencingAsyncResource');
-        return this[kEventEmitter];
-      }
-    }
-
-    EventEmitterAsyncResource =
-      class EventEmitterAsyncResource extends EventEmitter {
-        /**
-         * @param {{
-         *   name?: string,
-         *   triggerAsyncId?: number,
-         *   requireManualDestroy?: boolean,
-         * }} [options]
-         */
-        constructor(options = undefined) {
-          let name;
-          if (typeof options === 'string') {
-            name = options;
-            options = undefined;
-          } else {
-            if (new.target === EventEmitterAsyncResource) {
-              validateString(options?.name, 'options.name');
-            }
-            name = options?.name || new.target.name;
-          }
-          super(options);
-
-          this[kAsyncResource] =
-            new EventEmitterReferencingAsyncResource(this, name, options);
-        }
-
-        /**
-         * @param {symbol,string} event
-         * @param  {...any} args
-         * @returns {boolean}
-         */
-        emit(event, ...args) {
-          if (this[kAsyncResource] === undefined)
-            throw new ERR_INVALID_THIS('EventEmitterAsyncResource');
-          const { asyncResource } = this;
-          ArrayPrototypeUnshift(args, super.emit, this, event);
-          return ReflectApply(asyncResource.runInAsyncScope, asyncResource,
-                              args);
-        }
-
-        /**
-         * @returns {void}
-         */
-        emitDestroy() {
-          if (this[kAsyncResource] === undefined)
-            throw new ERR_INVALID_THIS('EventEmitterAsyncResource');
-          this.asyncResource.emitDestroy();
-        }
-
-        /**
-         * @type {number}
-         */
-        get asyncId() {
-          if (this[kAsyncResource] === undefined)
-            throw new ERR_INVALID_THIS('EventEmitterAsyncResource');
-          return this.asyncResource.asyncId();
-        }
-
-        /**
-         * @type {number}
-         */
-        get triggerAsyncId() {
-          if (this[kAsyncResource] === undefined)
-            throw new ERR_INVALID_THIS('EventEmitterAsyncResource');
-          return this.asyncResource.triggerAsyncId();
-        }
-
-        /**
-         * @type {EventEmitterReferencingAsyncResource}
-         */
-        get asyncResource() {
-          if (this[kAsyncResource] === undefined)
-            throw new ERR_INVALID_THIS('EventEmitterAsyncResource');
-          return this[kAsyncResource];
-        }
-      };
-  }
+  EventEmitterAsyncResource ??= require('internal/events/event_emitter_async_resource');
   return EventEmitterAsyncResource;
 }
